New packaging scheme deployed. This affects: * .c sources: commandline changes/fixes and config.h inclusion. * Makefile: huge changes to reflect the new packaging scheme. * all scripts: PATHES are no longer hardcoded. They follow the installation path set using ./configure. * other files: updated TODO, added some files like INSTALL. * eciconf.sh: fixed bugs introduced with last commit around the providers list.

+++ usermode/pusb-linux.c
@@ -359,7 +359,7 @@ int pusb_endpoint_rw(int fd,int ep,unsigned char * buf,int size,int timeout)
 		
 		received += ret;
     }
-	while (ret==bulk.len && size>0);
+	while (ret==(int)bulk.len && size>0);
 	
 	return received;
 }
